* unconstify vec3_origin

* remove sprites completly

* some cleaning

* move smd loading to kernel to not duplicate code for game modules

* server is aware of new model format

* fix two cases where md3s were assumed

* properly free server models, update `sv_modellist`, also fix `CS_CHEATS_ENABLED` not being updated in singleplayer

* delete annoying prints

* made export versions for sharing memory across modules
`Glob_HunkBegin, Glob_HunkAlloc, Glob_HunkEnd, Glob_HunkFree`

smdl: clip bone rotations according to smd specs

smdl: add surface names

* fix typo and export `Com_AnimationForName` to renderer

* bump renderer api version

* fix an oversight in mem

* initial x64 projects

* some fixes to x64

* flashlight hack

+++ src/qcommon/common.c
@@ -375,7 +375,7 @@ void MSG_WriteString (sizebuf_t *sb, char *s)
 	if (!s)
 		SZ_Write (sb, "", 1);
 	else
-		SZ_Write (sb, s, strlen(s)+1);
+		SZ_Write (sb, s, (int)strlen(s)+1);
 }
 
 void MSG_WriteCoord (sizebuf_t *sb, float f)
@@ -1082,7 +1082,7 @@ void SZ_Print (sizebuf_t *buf, char *data)
 {
 	int		len;
 	
-	len = strlen(data)+1;
+	len = (int)strlen(data)+1;
 
 	if (buf->cursize)
 	{
@@ -1194,7 +1194,7 @@ char *CopyString (char *in)
 {
 	char	*out;
 	
-	out = Z_Malloc (strlen(in)+1);
+	out = Z_Malloc ((int)strlen(in)+1);
 	strcpy (out, in);
 	return out;
 }
@@ -1376,7 +1376,7 @@ char* COM_NewString(char* string, memtag_t memtag)
 	char* newb, * new_p;
 	int		i, l;
 
-	l = strlen(string) + 1;
+	l = (int)strlen(string) + 1;
 
 	if (memtag > 0)
 		newb = Z_TagMalloc(l, memtag);
